shared/idmap: Split idmap
