+++ shared/idmap/utils.go
@@ -1,0 +1,5 @@
+package idmap
+
+func is_between(x, low, high int64) bool {
+	return x >= low && x < high
+}
